BLTI-172
The approach can be summarized as follows: 

(1) in a new window, if there is no error or message from the tool, the window silently closes, 
(2) in a new window is there is a message or error from the tool the message/error is shown along with a button to close the window, 
(3) in an iframe is there is a message or error it is shown (no button), 
(4) in an iframe if there is no message, a generic message is shown.

Hopefully the external tool does not start in an iframe and then take over top. Hopefully external tools that want to be outside of the iframe will open a new window. If it is a new window this work-flow works perfectly. 

This adds four new property strings to

basiclti-blis/src/bundle/blis.properties

outcome.tool.finished=External tool has finished
outcome.tool.lti_msg=External tool has finished:
outcome.tool.lti_errormsg=External tool has finished with an error:
outcome.tool.close.window=Continue...