SAK-22609 - Fix display of grouped Assignments for Gradebook
The AssignmentService was changed to check the current user's groups
for getAssignment in r98096 (apparently for SAK-21128). This requires
the grade info provider to apply an additional security advisor for
SECURE_ALL_GROUPS (asn.all.groups) in the assignment's site. The
provider could be simplified by extending the service API to include
an unauthenticated check for whether an assignment is grouped or not.

I have chosen to avoid changing the AssignmentService API for
compatibility and patch portability reasons, given that both the
interface and service implementation are very large and touched
relatively often.