SAK-15832 When directly visiting a joinable site allow the user to join.
This means if a user visits a site which they aren't currently a member of, but the site is joinable then they will be given the option to join the site. If they choose to they will be added to the site and then taken to the site they have just joined.

To disable this functionality set portal.redirectJoin=false in sakai.properties.