SAK-24208
Make it so that popups happen as a result of a side effect of a user-click action
whenever possible so we do not see the "Allow popups" dialogs on the various
browsers.

This was trickier than it seemed, touching /direct/site, three tools, the 
main portal, and PDA portal.

These modifications need to be committed or backported together as they have cross
dependencies.