SAK-25494
Refactor the PDA inlining code to properly implement the post-buffer check
for inlining (pre-buffer check was working fine all along).

With this change, profile2 seems to be properly inlining in the PDA portal 
so I changed the default of this property to be

portal.pda.iframesuppress=:all: 

We will need to carefully test these tools carefully in the PDA portal:

profile2, synoptic.messagecenter, sitestats, sitestats.admin                 

As these are now inlined in the PDA portal.